requirements.txt
Flask==3.1.0
qrcode==7.4.2
Pillow==10.3.0
python-dotenv==1.0.1
psycopg[binary]==3.2.3
